Move ArrowConverter into new core-arrow module
