+++ java/ingest/ingest-runner/pom.xml
@@ -131,6 +131,12 @@
         </dependency>
         <dependency>
             <groupId>sleeper</groupId>
+            <artifactId>core-arrow</artifactId>
+            <version>${project.parent.version}</version>
+            <scope>compile</scope>
+        </dependency>
+        <dependency>
+            <groupId>sleeper</groupId>
             <artifactId>statestore</artifactId>
             <version>${project.parent.version}</version>
             <type>test-jar</type>
